pyproject.toml
[tool.poetry]
name = "TelSök"
version = "0.1.0"
description = ""
authors = ["dunderrrrrr <emil.bjorkroth@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
htpy = "^24.6.1"
flask = "^3.0.3"
selenium = "^4.22.0"
google-api-python-client = "^2.136.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
